requirements.txt
python-telegram-bot[webhooks]==20.8
requests==2.31.0
qrcode==7.4.2
python-dotenv==1.0.0  # For local testing (optional)